+++ app/[locale]/(private)/leaves/page.tsx
@@ -1,70 +1,20 @@
 import { Pencil } from 'lucide-react';
 import { getTranslations } from 'next-intl/server';
-import { cookies } from 'next/headers';
-import { redirect } from 'next/navigation';
-
-import { SESSION_COOKIE } from '@/config';
 
 import { LeaveDeleteButton } from '@/app/_components/leave-delete-button';
 import { Button } from '@/app/_components/ui/button';
 import { Card, CardContent } from '@/app/_components/ui/card';
-import { getInjection } from '@/di/container';
+import { getLeavesForUser } from '@/app/actions/leaves';
 import { Link } from '@/i18n/navigation';
+import { getTotalLeaveDays } from '@/lib/leave';
 import { displayUKDateTime } from '@/lib/utils';
-import {
-  AuthenticationError,
-  UnauthenticatedError,
-} from '@/src/entities/errors/auth';
-
-async function getLeavesForUser() {
-  const instrumentationService = getInjection('IInstrumentationService');
-  return await instrumentationService.startSpan(
-    {
-      name: 'getLeavesForUser',
-      op: 'function.nextjs',
-    },
-    async () => {
-      try {
-        const getLeaveForUserController = getInjection(
-          'IGetLeavesForUserController'
-        );
-        const cookieStore = await cookies();
-        const token = cookieStore.get(SESSION_COOKIE)?.value;
-        const leaves = await getLeaveForUserController(token);
-        return leaves;
-      } catch (err) {
-        if (
-          err instanceof UnauthenticatedError ||
-          err instanceof AuthenticationError
-        ) {
-          redirect('/sign-in');
-        }
-        const crashReporterService = getInjection('ICrashReporterService');
-        crashReporterService.report(err);
-
-        throw err;
-      }
-    }
-  );
-}
-
-const getTotalLeaveDays = (leaves: { startDate: Date; endDate: Date }[]) => {
-  let result = 0;
-  for (let i = 0; i < leaves.length; i++) {
-    const { startDate, endDate } = leaves[i];
-    const t1 = startDate.getTime();
-    const t2 = endDate.getTime();
-    const days = Math.floor((t2 - t1) / (24 * 3600 * 1000));
-    result += days + 1;
-  }
-  return result;
-};
 
 export default async function LeavesPage() {
   const t = await getTranslations();
   let leaves;
   try {
-    leaves = await getLeavesForUser();
+    const res = await getLeavesForUser();
+    leaves = res.result;
   } catch (err) {
     throw err;
   }
@@ -75,7 +25,7 @@ export default async function LeavesPage() {
       data-cy="dashboard-content"
     >
       <div className="space-y-2">
-        {leaves.length === 0 && (
+        {leaves!.length === 0 && (
           <Card>
             <CardContent>
               <div className="flex flex-col justify-center items-center w-full">
@@ -84,7 +34,7 @@ export default async function LeavesPage() {
             </CardContent>
           </Card>
         )}
-        {leaves.map((leave) => {
+        {leaves!.map((leave) => {
           return (
             <div
               key={leave.id}
